+++ app/js/show-competitors.js
