+++ login.html
@@ -16,6 +16,6 @@
       <button id="buttonLogin" type="submit">Submit</button>
       <div id="loginError" style="display: none"></div>
     </dialog>
-    <script src="login.js"></script>
+    <script src="/app/js/login.js"></script>
   </body>
 </html>
